# Clone from existing shader instance to avoid extra cost.