If the fallback shader map is not empty, clone any existing shader instance in the map instead of acquiring via `MShaderManager::getFragmentShader()`, which creates new shader fragment graph for each shader instance and causes expensive shader compilation and rebinding.

This change will speed up the first frame loading time considerably.