@williamkrick 2020-03-13
I guess it doesn't really change the difficultly of varying the shaders more in the future.  If they have different parameters to set we'll have to handle that similar to color, and if the shaders really are different we'll have to update the search code to find a reasonable match to clone from.  But we can do all that when we need it.

We could add some debug code where in a debug build we clone from the map AND create one, and make sure they match.  That why if we do start varying the shaders more we'll immediately find this bit of code that needs to be updated.