@HdC-adsk 2020-03-13
Yes. Currently fallback shader is only assigned to USD objects without material assignment and the only variant for shading is displayColor/displayOpacity only. So we can leave it as is for now.

For those objects with material assignment we are also calling the `MShaderManager::getFragmentShader()` API but that part needs more thinking on how to share "shader" among shader instances if the material network is identical. So I leave that part unchanged for now.