@HdC-adsk 2020-03-18
@pilarmolinalopez Thanks for validation.

Although I do expect some speedup for the playback framerate, the major speedup from the change is for the first frame loading time, esp. for large assets. In some of our large assets, we see 145s to 35s speedup on slower machine and 90s to 20s speedup on faster machine. Did you observe similar speedup for the first frame? Thanks again.