@pilarmolinalopez 2020-03-18
hi @HdC-adsk - as you predicted, I'm seeing big speedup on first frame loading time:

- dev branch commit 00d651a (03/16/2020): First frame loading time = 12.63s
- this pr: First frame loading time = 5.96s